Permissions checks for IncreaseSpaceRequest
before, was possible for anybody to approve/decline their own request...

+++ ToTheMoon/ToTheMoon/Controllers/IncreaseSpaceRequestController.cs
@@ -116,7 +116,11 @@ namespace ToTheMoon.Controllers
             IncreaseSpaceRequest isr = db.IncreaseSpaceRequests.Find(increasespacerequest.ID);
             isr.brief = increasespacerequest.brief;
             isr.space = db.Spaces.Find(increasespacerequest.SpaceID);
-            if (ModelState.IsValid)
+
+            var manager = new UserManager<ApplicationUser>(new UserStore<ApplicationUser>(new ProjectContext()));
+            ApplicationUser currentUser = manager.FindById(User.Identity.GetUserId());
+
+            if (ModelState.IsValid && currentUser.role != GlobalRole.REGULAR)
             {
                 isr.timestamp = DateTime.Now;
                 db.SaveChanges();
@@ -175,14 +179,20 @@ namespace ToTheMoon.Controllers
         {
             IncreaseSpaceRequest incspacerequest = db.IncreaseSpaceRequests.Find(id);
 
-            db.IncreaseSpaceRequests.Remove(incspacerequest);
-            db.SaveChanges();
+            var manager = new UserManager<ApplicationUser>(new UserStore<ApplicationUser>(new ProjectContext()));
+            ApplicationUser currentUser = manager.FindById(User.Identity.GetUserId());
 
-            ///////////////////////////////
-            ///////////////////////////////
-            ////Send Email to Requester////
-            ///////////////////////////////
-            ///////////////////////////////
+            if(currentUser.role != GlobalRole.REGULAR)
+            {
+                db.IncreaseSpaceRequests.Remove(incspacerequest);
+                db.SaveChanges();
+
+                ///////////////////////////////
+                ///////////////////////////////
+                ////Send Email to Requester////
+                ///////////////////////////////
+                ///////////////////////////////
+            }
 
             return RedirectToAction("../Dashboard");
         }
